+++ CHANGELOG.md
@@ -85,4 +85,6 @@ Please add a _short_ line describing the PR you make, if the PR implements a spe
 * Updated token expiration information ([#245](https://github.com/scilifelabdatacentre/dds_cli/issues/245))
 
 ## Sprint (2022-02-23 - 2022-03-09)
+* Introduced a `--no-mail` flag in the CLI respectively a `send_email: True/False` json parameter to fix [issue 924](https://github.com/scilifelabdatacentre/dds_web/issues/924) ([#253](https://github.com/ScilifelabDataCentre/dds_cli/pull/253))
 * Added documentation and test protocol ([#252](https://github.com/ScilifelabDataCentre/dds_cli/pull/252))
+